Isolate /static and /proxy in REST API code
The purpose of these routes is just to make sure CORS works on IE9 with
help of the xdomain.js workaround. Since this code is very specific to
this odd browser and does not relate to the other code of the API, the
routes have been moved to the bottom of the file with a header comment
to explain their role. Hopefully we will remove them soon or later.