Move secret_key into separate restrited file
The secret_key used to sign tokens was stored as a simple parameter
REST API configuration file. This is always embarrasing to store such
sensitive among other parameters since it prevents from storing the
configuration file safely in SCM or so.

The secret key is now stored into a separate dedicated file. The
documentation has been update to explain how to generate a key file.

Fixes #108.