Add func to setup mocks before 1st req in REST API
This function plug_mocks() is run by Flask before serving the first
HTTP request thanks to the @before_first_request decorator.

Its purpose is to setup the mocks and fake/generated data when the
SLURM_WEB_CLUSTER_MOCK environment variable is set and the Flask app
runs in a development/testing environment.