Mock ConfigParser module in sys.modules
Since ConfigParser is used as soon as the REST API app is loaded, the
module must be mocked before the app is launched. Then the same
technique as for the LDAP and PySLURM  modules is used here, we
override the module namespace in sys.modules with our own fake module.

The fake module now implements exception and some missing methods.