Mock REST API at app load time
Previously, the REST API was eventually mocked before the HTTP first
request was server using Flask dedicated decorator. Unfortunately, this
technique does not work for all mocks since some external libraries are
used as soon as the app is loaded (eg. ConfigParser).

Now, the REST API is mocked as soon as the app is loaded, still when
the SLURM_WEB_CLUSTER_MOCK environment variable is set.