Show exact error in ajax error handler
The ajaxError jQuery handler considered all error as being
authentication error and reported them as such in the error #flash.
This is obviously not the case as many different kinds of errors
are possible. The new logic tries to report the exact error when
possible (ie. when not unexpected error with CORS), except on
/authentication route.

This commit also adds some hopefully useful comments to explain the
managed cases here.