Fix gantt time computatiton for jobs/QOS
When jobs/QOS are retrieved from the REST API, some jobs are not yet
started (typically PENDING) and have a start_time equal to 0. This
completely mess up the time computation done in computeTimesForJobs().
This happened to generate very (very) large times sets and cause out
of memory errors in the browser.

This commit simply adds a test to ignore jobs with a null start_time in
the logic.

Fix #122