Remove use of join in partitions view
A join function was used in the template of partition module to join
the array of nodes. This was actually a bug in PySLURM that has been
fixed in last versions (>= 16.05) with the following 2 commits:

- https://github.com/PySlurm/pyslurm/commit/a15f8b4dc5960c73a7b23077d04b1ebef7758881
- https://github.com/PySlurm/pyslurm/commit/c80a5a089bc7d376917de0601b48d991a24066d5

PySLURM now returns a string instead of an array of nodes. Therefore
the join function is not needed anymore, it can be safely removed from
slurm-web.

Fix #140