Show TRES instead of nodelist in jobs view
The 3rd column of the jobs view of the dashboard now shows the TRES of
the job instead of the nodelist. The nodelist could be long, quickly
truncated and therefore hard to read and often irrelevant to users.

Instead, the TRES gives a quick overview of the job size. The sorting
of this column is done on the number of cores.

The printNodes handlebars has been removed since it is not used
anymore.

Fix #89