Fix serial authentification failures
When the dashboard is configured with multiple cluster and AUTOLOGIN is
on, if the users try to login with a wrong password, their wrong
credentials were sent to all clusters' REST API. If all clusters
actually check authentifications credentials on the same LDAP server
(or replicates on this central LDAP server) and a security policy is
applied to lock accounts with several failed authentication tries, the
users may have their accounts locked on the LDAP server with only one
failed authentication attempt on the slurm-web dashboard.

Applying a series of action instead of parallel one solves the problem.
The error result object is different from the previous one thus the way
to obtain the error message has been changed accordingly.

Fix #137