Manage cache at PySLURM calls level
Up to now, the cache was managed at the API route call levels. This did
not offer great flexibility as it was impossible to share caches
between 2 routes requiring the same data from Slurm, typically jobs.

With this new function, cache is managed at function calls level. This
allows much finer grain control over the cache content and its usage.

This commit also removes the cache for nodeset route as it was wrong
and useless. The purpose of caching is to avoid on DOS on Slurm, that's
all.