Login begin with user's selection when autologin
With autologin on, the login procedure begins always from the first on
the list whether the user has selected another cluster to start with.
This prevents user from sucessfuly login if one only has access of
clusters which is not on the top of the list due to the modification
for #128. However, certains users will have to redo the login procedure
when trying to reach another cluster for the first time if none of
their granted clusters are on the top of the list. Further discussion
for a more intergrated plan will be needed.