Try to log in current cluster first
When autologin, each cluster is tested for connection, in the order they appear in window.clusters. But to prevent LDAP blacklisting the loop is exited upon first failure. Hence a user without access to the first window cluster can't log anywhere. Here we move the current cluster to the top of the list.