feat(gateway): asynchronous clusters list
Retrieve clusters stats and permissions in parallel with asynchronous
requests and aiohttp. Unfortunately, requests does not support
asynchronous requests in the same thread.

A new dependency is added on aiohttp package.

The module asyncrun is introduced to have an implementation of
asyncio.run() backported to Python 3.6. This alternative implementation
is only used on Python 3.6, more recent versions of Python use
implementation from the standard library.

fix #304