feat(gateway): possibility to hide denied clusters
Add the possibility to hide the list of denied clusters on which users
do not have any permission. New [ui]>hide_denied boolean gateway
configuration parameter is introduced to control activation of this
feature. By default, the boolean is false, ie. denied clusters are
visible, the behavior does not change.

fix #296