feat(front): HTTP/404 on message is not an error
The gateway API returns HTTP/404 if login service message is not
defined, then the error must not be reported in this case.

The GatewayDataGetter composable is extended to support custom error
handler. This is used by LoginServiceMessage component to ignore
HTTP/404 errors that could be returned by gateway API in absence of
the login service message. Other error are reported using default error
handler of GatewayDataGetter composable.