refactor(front): conv FormData.get() File to Blob
In Node JS, FormData.get() returns a File object instance when the part
is a file. This File is converted to a Blob so it can be loaded in
canvas. This is required to run this code in unit tests with Node JS and
it doesn't hurt browsers which don't need this additional conversion.