fix(front): add bearer token in racksdb request
Add missing bearer token in RacksDB infrastructure diagram request, it
is now required by @check_jwt decorator on the gateway endpoint with
RFL >= 1.3.0 to avoid HTTP/403 response.

The token is removed from request headers for login (incl. anonymous)
and login service message at the same time, as the client does not have
JWT at this stage yet and the gateway does not enforce the presence of
the bearer token for the corresponding endpoints.

fix #471