chore(ci): install importlib-metadata on py<3.8
On Python 3.8+, Slurm-web uses importlib.metadata in standard library to
retrieve the version of the package. For older versions, install
importlib_metadata external backport library, and explicitely force the
old version available in RHEL8.

Latest versions of markdown also uses importlib-metadata to retrieve the
list of installed markdown extensions. Unfortunately, it is not
compatible with importlib-metadata 0.23. For this reason, the version of
markdown package is also forced to match the version in RHEL8 (currently
2.6.11).