feat(agent): support JWT auto mode
Support slurmrestd JWT authentication auto mode in which Slurm-web agent
automatically generates and renews its token using the JWT signature key
shared with Slurm.

Slurm-web generates its token with a limited and configurable lifespan.
It can detect when the lifespan is about to be over and renews the token
when needed.