feat(gateway): replace base path placeholder
Replace at runtime base path placeholder in pre-built UI assets by URL
path extracted from configured UI public host to support running under
under a subfolder prefix on HTTP server

The UI prebuilt assets are first copied the source frontend UI assets to
a target directory, the placeholder base paths are then replaced in the
UI assets files in the target directory.

The target directory is created in the system's systemd runtime
directory, typically /run/slurm-web-gateway/ui, or in a temporary
directory if the RUNTIME_DIRECTORY environment variable is not set.

The placeholder base path, set at frontend application build time, is
replaced with the path configured in the public URL prefix. This makes
the same build usable under any path without rebuilding the frontend
application.

This approach is not great, to say the least, but considering the limits
of the frontend application build process, it's the best we have found
so far.

fix #388