fix(front): skip rendering empty job account
The association.account field is empty in Slurm REST API v0.0.43 and
v0.0.44. This is probably a regression in Slurm:

https://support.schedmd.com/show_bug.cgi?id=24215

This leads to an error when the JobFieldRaw is rendering with an empty
query parameter for account link. Then skip rendering of JobFieldRaw
component on empty account as a workaround.